@@ -1,3 +1,3 @@
 module badstuff
 
-go 1.21.6
+go 1.22.0
